+++ tests/test_modes.py
@@ -35,8 +35,8 @@ def test_modes(mode, problem_create):
     )
     log_filepath = sorted(glob.glob(log_filepath), reverse=True)[-1]
     with open(log_filepath, "r") as F:
-        last_line = F.readlines()[-2]
-    assert "INFO : Process completed." in last_line
+        last_line = F.readlines()[-2:]
+    assert "INFO : Process completed." in last_line[0] or last_line[1]
 
 
 # Test to check if the outputs are the same as expected outcomes
@@ -157,7 +157,7 @@ def test_omic_datasets(mode, omic, problem):
         config["data"]["save_path"][1:]
         + f'results/{config["data"]["name"]}/AutoOmicLog_*'
     )
-    log_filepath = glob.glob(log_filepath)[-1]
+    log_filepath = sorted(glob.glob(log_filepath), reverse=True)[-1]
     with open(log_filepath, "r") as F:
-        last_line = F.readlines()[-1]
-    assert "INFO : Process completed." in last_line
+        last_line = F.readlines()[-2:]
+    assert "INFO : Process completed." in last_line[0] or last_line[1]